bugfix: enable ExtraWarnings for the Nao robot explicitely

+++ Framework/NaoTH-Commons/Source/Tools/Math/Matrix_nxn.h
@@ -382,7 +382,7 @@ public:
     // initialize ranking vector
     Vector_n<int, N> ranking;
     int i;
-    for (i = 0; i < N; ++i)
+    for (i = 0; i < (int)N; ++i)
       ranking[i] = i;
     
     T z = T();
